refactor: Race default constructor,race directory
