+++ 고형인/java-racingcar-6/src/main/java/racingcar/Application.java
@@ -1,5 +1,6 @@
 package racingcar;
 
+import racingcar.race.Race;
 import racingcar.service.InputService;
 import racingcar.serviceimpl.CarNameInputServiceImpl;
 import racingcar.serviceimpl.CountInputServiceImpl;
